feat(rust, python): support negative index in `pct_change`

+++ polars/polars-lazy/polars-plan/src/dsl/function_expr/mod.rs
@@ -113,7 +113,7 @@ pub enum FunctionExpr {
     Coalesce,
     ShrinkType,
     #[cfg(feature = "diff")]
-    Diff(usize, NullBehavior),
+    Diff(i64, NullBehavior),
     #[cfg(feature = "interpolate")]
     Interpolate(InterpolationMethod),
     #[cfg(feature = "dot_product")]
